+++ loaders/chains.go
@@ -55,6 +55,24 @@ func LoadChainDefinition(chainName string, definition ...string) (*definitions.C
 	return chain, nil
 }
 
+func LoadChainTypes(fileName string) (*definitions.ChainType, error) {
+	fileName = filepath.Join(config.ChainTypePath, fileName)
+	log.WithField("file name", fileName).Info("Loading Chain Definition.")
+	var typ = definitions.BlankChainType()
+	var chainType = viper.New()
+
+	if err := getSetup(fileName, chainType); err != nil {
+		return nil, err
+	}
+
+	// marshall file
+	if err := chainType.Unmarshal(typ); err != nil {
+		return nil, fmt.Errorf("\nSorry, the marmots could not figure that chain types file out.\nPlease check your chain type definition file is properly formatted.\nERROR =>\t\t\t%v", err)
+	}
+
+	return typ, nil
+}
+
 // ChainsAsAService convert the chain definition to a service one
 // and set the CHAIN_ID environment variable. ChainsAsAService
 // can return config load errors.
