+++ maker/files.go
@@ -1,4 +1,4 @@
-package util
+package maker
 
 import (
 	"encoding/json"
